refactor the bulk export API into its own target
