+++ Sources/SpeziHealthKit/SpeziHealthKit.docc/SpeziHealthKit.md
@@ -88,9 +88,6 @@ class ExampleAppDelegate: SpeziAppDelegate {
 - ``AnySampleType``
 - ``SampleTypesCollection``
 
-### Bulk Exporting of Historical Health Data
-- ``BulkHealthExporter``
-
 ### Working with ECG Data
 - ``HealthKit/HKElectrocardiogram/symptoms(from:)``
 - ``HealthKit/HKElectrocardiogram/voltageMeasurements(from:)``
